fix build

+++ CRC8.h
@@ -35,7 +35,7 @@ public:
   bool     getReverseOut() { return _reverseOut; };
 
   void     add(uint8_t value);
-  void     add(const uint8_t * array, uint8_t length);
+  void     add(const uint8_t * array, uint16_t length);
 
   uint8_t  getCRC();  // returns CRC
   uint32_t count()    { return _count; };
